Use http port to validate grpc exporter, unify mocked_servers images into one repo with different tags
