--- mocked_server/grpc_metrics/metrichandler.proto
@@ -1,23 +1,0 @@
-syntax = "proto3";
-
-package opentelemetry.proto.collector.metrics.v1;
-
-// The Metric Handler service definition.
-service MetricsService {
-    // Export data
-    rpc Export (Request) returns (ExportMetricsResponse) {}
-    // Check data
-    rpc CheckData (Request) returns (CheckDataResponse) {}
-}
-
-message Request {
-    string name = 1;
-}
-
-message ExportMetricsResponse {
-    string message = 1;
-}
-
-message CheckDataResponse {
-    string message = 1;
-}
