+++ mocked_servers/grpc_metrics/Dockerfile
