+++ mocked_servers/grpc_trace/Dockerfile
