+++ mocked_servers/grpc_trace/package.json
@@ -12,6 +12,7 @@
     "@grpc/grpc-js": "^1.1.0",
     "@grpc/proto-loader": "^0.5.0",
     "async": "^1.5.2",
+    "express": "^4.17.1",
     "google-protobuf": "^3.0.0"
   }
 }
