+++ mocked_servers/https/Dockerfile
@@ -2,7 +2,7 @@ FROM node:14
 
 WORKDIR /app
 
-COPY ./certificates ./certificates
+COPY certificates ./certificates
 COPY *.js ./
 COPY package.json ./
