+++ mocked_servers/https/certificates/ssl/ca-bundle.crt
