+++ terraform/canary/main.tf
@@ -28,6 +28,7 @@ module "basic_components" {
   testing_id = module.common.testing_id
   mocked_endpoint = var.mock_endpoint
   sample_app = var.sample_app
+  mocked_server = var.mocked_server
 }
 
 # launch ec2
