+++ terraform/common.tf
@@ -20,6 +20,10 @@ variable "mocked_server_image" {
   default = ""
 }
 
+variable "mocked_server" {
+  default = "https"
+}
+
 # we use sample_app_image if it's not empty, if it's empty, the sample_app_image will come from the basic component, which is built by imagebuild module
 # instead "sample_app" will be used to choose the image
 variable "sample_app_image" {
