+++ terraform/ecs/variables.tf
@@ -21,6 +21,6 @@ variable "sample_app_callable" {
   default = true
 }
 
-variable "mocked_server_validating_url_type" {
-  default = ""
+variable "mock_endpoint" {
+  default = "localhost/put-data"
 }
