+++ terraform/testcases/otlp_grpc_exporter_trace_mock/parameters.tfvars
@@ -1,4 +1,4 @@
 mock_endpoint = "mocked-server:55670"
-mocked_server_validating_url_type = "grpc"
+mocked_server = "grpc_trace"
 # data type will be emitted. Possible values: metric or trace
 soaking_data_mode = "trace"
