--- validator/src/main/java/com/amazon/aoc/clients/GrpcClient.java
@@ -1,5 +1,0 @@
-package com.amazon.aoc.clients;
-
-public interface GrpcClient {
-  String checkData(String name);
-}
