+++ validator/src/main/java/com/amazon/aoc/models/Context.java
@@ -34,8 +34,6 @@ public class Context {
 
   private ECSContext ecsContext;
 
-  private String testcase;
-
   /*
   alarm related parameters
    */
