+++ validator/src/main/java/com/amazon/aoc/validators/MockedServerValidator.java
@@ -1,31 +1,21 @@
 package com.amazon.aoc.validators;
 
 import com.amazon.aoc.callers.ICaller;
-import com.amazon.aoc.clients.GrpcClient;
-import com.amazon.aoc.clients.GrpcClientFactory;
-import com.amazon.aoc.clients.GrpcMetricsClient;
-import com.amazon.aoc.clients.GrpcTraceClient;
 import com.amazon.aoc.exception.BaseException;
 import com.amazon.aoc.exception.ExceptionCode;
 import com.amazon.aoc.fileconfigs.FileConfig;
 import com.amazon.aoc.helpers.RetryHelper;
 import com.amazon.aoc.models.Context;
 import com.amazon.aoc.models.ValidationConfig;
-import io.grpc.ManagedChannel;
-import io.grpc.ManagedChannelBuilder;
 import lombok.extern.log4j.Log4j2;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 
-import java.util.concurrent.TimeUnit;
-
 @Log4j2
 public class MockedServerValidator implements IValidator {
   String mockedServerValidatingUrl;
   ICaller caller;
-  String testcase;
-  private static final String HTTP_IDENTIFIER = "http";
 
   @Override
   public void init(
@@ -36,7 +26,6 @@ public class MockedServerValidator implements IValidator {
       throws Exception {
     mockedServerValidatingUrl = context.getMockedServerValidatingUrl();
     this.caller = caller;
-    this.testcase = context.getTestcase();
   }
 
   @Override
@@ -47,14 +36,10 @@ public class MockedServerValidator implements IValidator {
     }
 
     // hit the mocked server to validate to if it receives data
-    if (isHttpServer()) {
-      callHttpMockedServer();
-    } else {
-      callGrpcMockedServer();
-    }
+    callMockedServer();
   }
 
-  private void callHttpMockedServer() throws Exception {
+  private void callMockedServer() throws Exception {
     OkHttpClient client = new OkHttpClient();
     Request request = new Request.Builder().url(this.mockedServerValidatingUrl).build();
 
@@ -70,36 +55,7 @@ public class MockedServerValidator implements IValidator {
             throw new BaseException(ExceptionCode.MOCKED_SERVER_NOT_RECEIVE_DATA);
           }
 
-          log.info("http mocked server validation passed");
+          log.info("mocked server validation passed");
         });
-  }
-
-  private void callGrpcMockedServer() throws Exception {
-    ManagedChannel channel = ManagedChannelBuilder.forTarget(this.mockedServerValidatingUrl)
-            .usePlaintext()
-            .build();
-    GrpcClient client = new GrpcClientFactory().getGrpcClient(this.testcase, channel);
-
-    try {
-      RetryHelper.retry(
-          () -> {
-            String response = client.checkData("");
-            if (null == response) {
-              throw new BaseException(ExceptionCode.MOCKED_SERVER_NOT_AVAILABLE);
-            }
-
-            if (!response.equalsIgnoreCase("success")) {
-              throw new BaseException(ExceptionCode.MOCKED_SERVER_NOT_RECEIVE_DATA);
-            }
-
-            log.info("grpc mocked server validation passed");
-          });
-    } finally {
-      channel.shutdownNow().awaitTermination(5, TimeUnit.SECONDS);
-    }
-  }
-
-  private boolean isHttpServer() {
-    return this.mockedServerValidatingUrl.startsWith(HTTP_IDENTIFIER);
   }
 }
