--- validator/src/main/proto/tracehandler.proto
@@ -1,25 +1,0 @@
-syntax = "proto3";
-
-package opentelemetry.proto.collector.trace.v1;
-
-option java_package = "com.amazon.aoc.clients";
-
-// The Trace Handler service definition.
-service TraceService {
-  // Export data
-  rpc Export (Request) returns (ExportTraceServiceResponse) {}
-  // Check data
-  rpc CheckData (Request) returns (CheckDataResponse) {}
-}
-
-message Request {
-  string name = 1;
-}
-
-message ExportTraceServiceResponse {
-  string message = 1;
-}
-
-message CheckDataResponse {
-  string message = 1;
-}
